# backend/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "GestorPlus API",
        "description": "Documentación Swagger para la API de GestorPlus",
        "contact": {
            "name": "Soporte GestorPlus",
            "email": "soporte@tudominio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/",
            "description": "Servidor local"
        },
        {
            "url": "https://api.tudominio.com/",
            "description": "Servidor producción"
        }
    ],
    "paths": {
...
        "/jornadas/corroborar": {
            "patch": {
                "tags": [
                    "Jornada"
                ],
                "summary": "Corroborar jornada",
                "description": "Corrobora una jornada por idJornada.",
                "operationId": "d08e1879c0634c729b4ba6609580b367",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idJornada"
                                ],
                                "properties": {
                                    "idJornada": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jornada corroborada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "JornadaCorroborada": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Faltan parámetros requeridos"
                    }
                }
            }
        },
        "/jornadas/no-corroborar": {
            "patch": {
                "tags": [
                    "Jornada"
                ],
                "summary": "No corroborar jornada",
                "description": "Marca una jornada como no corroborada por idJornada.",
                "operationId": "c3f533899d28b9345ecb569ee1749490",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "idJornada"
                                ],
                                "properties": {
                                    "idJornada": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jornada no corroborada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "JornadaNoCorroborada": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Faltan parámetros requeridos"
                    }
                }
            }
        },
        "/jornadas/finalizar": {
            "patch": {
                "tags": [
                    "Jornada"
                ],
                "summary": "Finalizar jornada",
                "description": "Finaliza una jornada por fechaBogota (en el body) y usuario autenticado (token).",
                "operationId": "1aa4160eec8b42296f180fc7989bc1f6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "fechaBogota"
                                ],
                                "properties": {
                                    "fechaBogota": {
                                        "type": "string",
                                        "example": "2024-06-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jornada finalizada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "JornadaFinalizada": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Faltan parámetros requeridos"
                    },
                    "500": {
                        "description": "Error al finalizar la jornada"
                    }
                }
            }
        },
...
    },
...
}